requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
pyspellchecker
python-dotenv
groq
